NeoConnection doesn't support self-signed certs
Add a new configuration property to NeoConnection (and matching checkbox
to the NeoConnectionDialog ui) for toggling a trust strategy for
trusting all certificates. This allows users to easily mimic using the
bolt+ssc:// and neo4j+ssc:// schemes.

This does NOT add support for the more complicated trust strategy of
providing a signature File to determine trust.

Tested manually with both Neo4j 3.5, 4.1, 4.2, and Aura.